+++ R/dists.R
@@ -2,4 +2,4 @@
 
 survstan_distributions <- c("exponential", "weibull", "lognormal", "loglogistic",
                             "fatigue", "gamma", "rayleigh", "gompertz",
-                            "ggstacy", "ggprentice")
+                            "ggstacy", "ggprentice", "bernstein", "piecewise")