+++ R/survfit.R
@@ -1,35 +1,35 @@
 
-surv_aft <- function(time, pars, lp, baseline, p){
+surv_aft <- function(time, pars, lp, baseline, p, m){
   time <- time*exp(-lp)
-  H0 <- cumhaz(time, pars, baseline, p)
+  H0 <- cumhaz(time, pars, baseline, p, m)
   surv <- exp(-H0)
   return(surv)
 }
 
-surv_ah <- function(time, pars, lp, baseline, p){
+surv_ah <- function(time, pars, lp, baseline, p, m){
   time <- time*exp(-lp)
-  H0 <- cumhaz(time, pars, baseline, p)
+  H0 <- cumhaz(time, pars, baseline, p, m)
   surv <- exp(-H0*exp(lp))
   return(surv)
 }
 
-surv_ph <- function(time, pars, lp, baseline, p){
-  H0 <- cumhaz(time, pars, baseline, p)
+surv_ph <- function(time, pars, lp, baseline, p, m, rho){
+  H0 <- cumhaz(time, pars, baseline, p, m)
   surv <- exp(-H0*exp(lp))
   return(surv)
 }
 
-surv_po <- function(time, pars, lp, baseline, p){
-  H0 <- cumhaz(time, pars, baseline, p)
+surv_po <- function(time, pars, lp, baseline, p, m, rho){
+  H0 <- cumhaz(time, pars, baseline, p, m, rho)
   Rt = expm1(H0)*exp(lp)
   surv <- exp(-log1p(Rt))
   return(surv)
 }
 
 
-surv_yp <- function(time, pars, lp_short, lp_long, baseline, p){
+surv_yp <- function(time, pars, lp_short, lp_long, baseline, p, m, rho){
   p <- 2*p
-  H0 <- cumhaz(time, pars, baseline, p)
+  H0 <- cumhaz(time, pars, baseline, p, m, rho)
   ratio <- exp(lp_short - lp_long)
   Rt = expm1(H0)*ratio
   theta <- exp(lp_long)
@@ -38,10 +38,10 @@ surv_yp <- function(time, pars, lp_short, lp_long, baseline, p){
 }
 
 
-surv_eh <- function(time, pars, lp1, lp2, baseline, p){
+surv_eh <- function(time, pars, lp1, lp2, baseline, p, m){
   p <- 2*p
   time <- time/exp(lp1)
-  H0 <- cumhaz(time, pars, baseline, p)
+  H0 <- cumhaz(time, pars, baseline, p, m)
   surv <- exp(- H0*exp(lp1+lp2))
   return(surv)
 }
@@ -80,9 +80,11 @@ survfit.survstan <- function(formula, newdata = NULL, ...){
   labels <- names(mf)[-1]
   time <- c(0,sort( stats::model.response(mf)[,1]))
   pars <- estimates(object)
+  m <- object$m
+  rho <- object$rho
 
   if(is.null(newdata)){
-    H0 <- cumhaz(time, pars, baseline, p=0)
+    H0 <- cumhaz(time, pars, baseline, p=0, m, rho)
     surv <- data.frame(
       time = time,
       surv = exp(-H0)
@@ -122,12 +124,12 @@ survfit.survstan <- function(formula, newdata = NULL, ...){
     J <- N/n
 
     surv <- switch (survreg,
-                    "aft" = with(df, surv_aft(time, pars, lp, baseline, p)),
-                    "ph" = with(df, surv_ph(time, pars, lp, baseline, p)),
-                    "po" = with(df, surv_po(time, pars, lp, baseline, p)),
-                    "ah" = with(df, surv_ah(time, pars, lp, baseline, p)),
-                    "yp" = with(df, surv_yp(time, pars, lp_short, lp_long, baseline, p)),
-                    "eh" = with(df, surv_eh(time, pars, lp1, lp2, baseline, p))
+                    "aft" = with(df, surv_aft(time, pars, lp, baseline, p, m)),
+                    "ph" = with(df, surv_ph(time, pars, lp, baseline, p, m, rho)),
+                    "po" = with(df, surv_po(time, pars, lp, baseline, p, m, rho)),
+                    "ah" = with(df, surv_ah(time, pars, lp, baseline, p, m)),
+                    "yp" = with(df, surv_yp(time, pars, lp_short, lp_long, baseline, p, m, rho)),
+                    "eh" = with(df, surv_eh(time, pars, lp1, lp2, baseline, p, m))
     )
 
     surv <- df %>%